Refactored multi-class/enum file into separate files.
