+++ src/Enums/VideoCodecsEnum.cs
@@ -1,0 +1,34 @@
+﻿//-----------------------------------------------------------------------------
+// Filename: VideoCodecsEnum.cs
+//
+// Description: Enum for common video codecs.
+//
+// Author(s):
+// Aaron Clauson (aaron@sipsorcery.com)
+// 
+// History:
+// 20 May 2025  Aaron Clauson   Refactored from MediaEndPoints.
+//
+// License: 
+// BSD 3-Clause "New" or "Revised" License and the additional
+// BDS BY-NC-SA restriction, see included LICENSE.md file.
+//-----------------------------------------------------------------------------
+
+namespace SIPSorceryMedia.Abstractions;
+
+public enum VideoCodecsEnum
+{
+    CELB,
+    JPEG,
+    NV,
+    H261,
+    MPV,
+    MP2T,
+    H263,
+    VP8,
+    VP9,
+    H264,
+    H265,
+
+    Unknown
+}
